Refactor hasTabIndex on Element
