+++ packages/alfa-dom/src/node/element/predicate.ts
@@ -1,4 +1,5 @@
 export * from "./predicate/has-id";
 export * from "./predicate/has-name";
 export * from "./predicate/has-namespace";
+export * from "./predicate/has-tab-index";
 export * from "./predicate/is-disabled";
