+++ packages/alfa-rules/tsconfig.json
@@ -20,7 +20,6 @@
     "src/common/predicate/has-name-from.ts",
     "src/common/predicate/has-nondefault-role.ts",
     "src/common/predicate/has-role.ts",
-    "src/common/predicate/has-tab-index.ts",
     "src/common/predicate/has-text-content.ts",
     "src/common/predicate/has-unique-id.ts",
     "src/common/predicate/has-value.ts",
